refactor(builder): relocate dead identity dispatcher to _deprecated/
Moves 7 files now unreachable from production into a _deprecated/
subfolder, making it visually obvious in the file tree that the legacy
ResizeObserver-based dispatcher and its three identity-layout variants
are no longer wired in.

Files moved (history preserved via git mv):
- lanes/active-row.tsx
- lanes/identity/identity-lane.tsx (the dispatcher)
- lanes/identity/identity-single-row.tsx
- lanes/identity/identity-mid-stack.tsx
- lanes/identity/identity-vertical.tsx
- lanes/identity/identity-lane-ghost.tsx
- use-container-compact.ts

Plus the lanes/identity/ directory is removed (its index.ts barrel
deleted, no remaining contents). 4 test files updated to point at the
new paths; the tests still exercise the legacy components for behavior
documentation purposes (and to make later deletion safe).

Tests: 90 suites, 1963 tests pass. Typecheck clean.
