+++ apps/web/src/components/team-builder/_deprecated/identity-lane.tsx
@@ -5,11 +5,11 @@ import { useRef } from "react";
 import { type GameFormat } from "@trainers/pokemon";
 import { type Tables, type TablesUpdate } from "@trainers/supabase";
 
-import { type ValidationError } from "../../validation-hooks";
-import { useContainerCompact } from "../../use-container-compact";
-import { useTeamLayoutMode } from "../../use-team-layout";
+import { type ValidationError } from "../validation-hooks";
+import { useContainerCompact } from "./use-container-compact";
+import { useTeamLayoutMode } from "../use-team-layout";
 import { IdentityLaneGhost } from "./identity-lane-ghost";
-import { filterCurrentTeam } from "../../shared/identity-layout-props";
+import { filterCurrentTeam } from "../shared/identity-layout-props";
 import { IdentityMidStack } from "./identity-mid-stack";
 import { IdentitySingleRow } from "./identity-single-row";
 import { IdentityVertical } from "./identity-vertical";
