--- apps/web/src/components/team-builder/lanes/identity/index.ts
@@ -1,1 +1,0 @@
-export { IdentityLane } from "./identity-lane";
